requirements.txt
streamlit>=1.35.0
pillow
passlib[bcrypt]
authlib>=1.3.0
requests>=2.31.0
